chore: break out schema to files
