+++ apps/course-builder-web/src/db/mysql-table.ts
@@ -1,0 +1,11 @@
+import { mysqlTableCreator } from 'drizzle-orm/mysql-core'
+
+/**
+ * This is an example of how to use the multi-project schema feature of Drizzle ORM. Use the same
+ * database instance for multiple projects.
+ *
+ * @see https://orm.drizzle.team/docs/goodies#multi-project-schema
+ */
+export const mysqlTable = mysqlTableCreator(
+	(name) => `${process.env.NEXT_PUBLIC_APP_NAME}_${name}`,
+)
