+++ apps/course-builder-web/src/db/schemas/role-permissions.ts
@@ -1,0 +1,34 @@
+import { mysqlTable } from '@/db/mysql-table'
+import {
+	boolean,
+	index,
+	primaryKey,
+	timestamp,
+	varchar,
+} from 'drizzle-orm/mysql-core'
+
+export const rolePermissions = mysqlTable(
+	'rolePermission',
+	{
+		roleId: varchar('roleId', { length: 255 }).notNull(),
+		permissionId: varchar('permissionId', { length: 255 }).notNull(),
+		active: boolean('active').notNull().default(true),
+		createdAt: timestamp('createdAt', {
+			mode: 'date',
+			fsp: 3,
+		}).defaultNow(),
+		updatedAt: timestamp('updatedAt', {
+			mode: 'date',
+			fsp: 3,
+		}).defaultNow(),
+		deletedAt: timestamp('deletedAt', {
+			mode: 'date',
+			fsp: 3,
+		}),
+	},
+	(rp) => ({
+		pk: primaryKey({ columns: [rp.roleId, rp.permissionId] }),
+		roleIdIdx: index('roleId_idx').on(rp.roleId),
+		permissionIdIdx: index('permissionId_idx').on(rp.permissionId),
+	}),
+)
