+++ apps/course-builder-web/src/db/schemas/sessions.ts
@@ -1,0 +1,16 @@
+import { mysqlTable } from '@/db/mysql-table'
+import { index, timestamp, varchar } from 'drizzle-orm/mysql-core'
+
+export const sessions = mysqlTable(
+	'session',
+	{
+		sessionToken: varchar('sessionToken', { length: 255 })
+			.notNull()
+			.primaryKey(),
+		userId: varchar('userId', { length: 255 }).notNull(),
+		expires: timestamp('expires', { mode: 'date' }).notNull(),
+	},
+	(session) => ({
+		userIdIdx: index('userId_idx').on(session.userId),
+	}),
+)
